float

+++ eos_tax/models.py
@@ -19,8 +19,8 @@ class MonthlyTax(models.Model):
     month = models.IntegerField(verbose_name="Taxed month", blank=False, default=0)
     year = models.IntegerField(verbose_name="Taxed year", blank=False, default=0)
     corp_name = models.CharField(verbose_name="Corporation name", max_length=254, blank=True, default='')
-    tax_value = models.BigIntegerField (verbose_name="Tax value", blank=False, default=0)
-    tax_percentage = models.IntegerField(verbose_name="Tax percentage", blank=False, default=0)
+    tax_value = models.BigIntegerField(verbose_name="Tax value", blank=False, default=0)
+    tax_percentage = models.FloatField(verbose_name="Tax percentage", blank=False, default=0)
     payed = models.BooleanField(verbose_name="Payed", default=False)
